pyproject.toml
[project]
name = "career-mentor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.5",
    "openai-agents[litellm]>=0.2.4",
    "python-dotenv>=1.1.1",
]
